Move hash_random into its own utility module.
